fix: enable HDF5, Julia, and Boost tests on macOS ARM64
- Fix size_t vs uint64_t mismatches (different types on ARM macOS):
  hdf5N2Write, fakeGpu, gpuSimulateRFISK
- Replace deprecated std::result_of_t with std::invoke_result_t in
  juliaManager.hpp (deprecated in C++17, errors on newer Apple Clang)
- Guard -latomic link flag with NOT APPLE (built into compiler runtime)
- Remove boost_system from required Boost components (header-only
  since Boost 1.69)
- Add macOS-compatible linker flags for test_Telescope and test_N2Subset
- Add global LAPACKE include path so Blaze can find cblas.h in tests

+++ lib/testing/fakeGpu.cpp
@@ -227,7 +227,7 @@ uint64_t FakeTelescope::to_seq(timespec time) const {
     return ns / seq_length_nsec();
 }
 
-size_t FakeTelescope::seq_length_nsec() const {
+uint64_t FakeTelescope::seq_length_nsec() const {
     return 2560;
 }
